refactored and added the usertoken
