+++ api/Data/TheBattleLibrary.Data.Entities/IApplicationDbContext.cs
@@ -1,18 +1,18 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
-using TheBattleLibrary.Data.Entities;
 
-namespace TheBattleLibrary.Data
+namespace TheBattleLibrary.Data.Entities;
+
+public interface IApplicationDbContext
 {
-    public interface IApplicationDbContext
-    {
-        DbSet<UserAccount> Users { get; set; }
-        
-        DbSet<BattleList> BattleLists { get; set; } 
-        DbSet<Selection> Selections { get; set; }
-        
-        DatabaseFacade Database { get; }
+    DbSet<UserAccount> Users { get; set; }
+    
+    DbSet<BattleList> BattleLists { get; set; } 
+    DbSet<Selection> Selections { get; set; }
+    
+    DbSet<UserToken> UserTokens { get; set; }
+    
+    DatabaseFacade Database { get; }
 
-        int SaveChanges();
-    }
+    int SaveChanges();
 }
