+++ api/Data/TheBattleLibrary.Data.Sqlite/SqliteDbContext.cs
@@ -1,15 +1,14 @@
 ﻿using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Design;
-using Microsoft.EntityFrameworkCore.Infrastructure;
 using TheBattleLibrary.Data.Entities;
 
-namespace TheBattleLibrary.Data;
+namespace TheBattleLibrary.Data.Sqlite;
 
 public class SqliteDbContext : DbContext, IApplicationDbContext
 {
     public DbSet<UserAccount> Users { get; set; }
     public DbSet<BattleList> BattleLists { get; set; }
     public DbSet<Selection> Selections { get; set; }
+    public DbSet<UserToken> UserTokens { get; set; }
 
     public SqliteDbContext(DbContextOptions<SqliteDbContext> options) : base(options)
     {
@@ -30,14 +29,4 @@ public class SqliteDbContext : DbContext, IApplicationDbContext
         base.OnModelCreating(modelBuilder);
     }
 
-}
-
-public class DesignTimeDbContextFactory : IDesignTimeDbContextFactory<SqliteDbContext>
-{
-    public SqliteDbContext CreateDbContext(string[] args)
-    {
-        var optionsBuilder = new DbContextOptionsBuilder<SqliteDbContext>(); 
-        optionsBuilder.UseSqlite("Data Source=your_database.db;"); 
-        return new SqliteDbContext(optionsBuilder.Options);
-    }
 }
