+++ api/Data/TheBattleLibrary.Data.Tests/StartupTests.cs
@@ -5,6 +5,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.EntityFrameworkCore;
 using System;
+using TheBattleLibrary.Data.Entities;
 
 namespace TheBattleLibrary.Data.Tests;
