+++ api/Data/TheBattleLibrary.Data/Startup.cs
@@ -3,6 +3,8 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 using System;
+using TheBattleLibrary.Data.Entities;
+using TheBattleLibrary.Data.Sqlite;
 
 namespace TheBattleLibrary.Data;
